+++ server/src/main/java/org/opensearch/node/Node.java
@@ -902,7 +902,8 @@ public class Node implements Closeable {
                 systemIndices,
                 forbidPrivateIndexSettings,
                 awarenessReplicaBalance,
-                remoteStoreSettings
+                remoteStoreSettings,
+                repositoriesServiceReference::get
             );
             pluginsService.filterPlugins(Plugin.class)
                 .forEach(