+++ server/src/test/java/org/opensearch/snapshots/SnapshotResiliencyTests.java
@@ -2171,7 +2171,8 @@ public class SnapshotResiliencyTests extends OpenSearchTestCase {
                     systemIndices,
                     false,
                     new AwarenessReplicaBalance(Settings.EMPTY, clusterService.getClusterSettings()),
-                    DefaultRemoteStoreSettings.INSTANCE
+                    DefaultRemoteStoreSettings.INSTANCE,
+                    null
                 );
                 actions.put(
                     CreateIndexAction.INSTANCE,